move minimal jsb init into a separate header
