+++ cocos/bindings/manual/jsb_cocos_manual.cpp
@@ -27,7 +27,7 @@
 
 #include "cocos/bindings/jswrapper/SeApi.h"
 #include "cocos/bindings/manual/jsb_conversions.h"
-#include "cocos/bindings/manual/jsb_global.h"
+#include "cocos/bindings/manual/jsb_global_init.h"
 #include "cocos/bindings/auto/jsb_cocos_auto.h"
 
 #include "storage/local-storage/LocalStorage.h"
